Moved version into package init
